requirements.txt
<<<<<<< HEAD
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions==1.24.0
azure-storage-blob==12.16.0
=======
azure-functions==1.24.0
azure-storage-blob==12.16.0
>>>>>>> 61d03e4a51399d3498ee7cb7616c93289f66e51f
